pyproject.toml
[project]
name = "viscomp-back"
version = "0.1.0"
description = "FastAPI backend for viscomp"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "python-multipart>=0.0.9",
    "transformers>=4.37.2",
    "torch>=2.2.0",
    "Pillow>=10.2.0",
    "numpy>=1.26.0"
]
